Fix a crash when extracting test names from XCTest binary.
The code that extracted the swift tests was fragile in that if a
function that is not a test just happened to have the `_test` string,
it would get pulled out and not match the expected format.

This commit makes the testing a bit more robust by expecting exaclty
three fields in the swift tests. It also fixes a potential bug in
which objective C tests could be missed if there were other functions
in the middle (I am not sure this was actually happening but it
could).