Fix running of tests that use OCMock
This is a bit of a hack but it fixes a problem where if a
test (layout) uses OCMock, the mocks won't be fired. I have a
hypothesis that some framework that's loaded before the OCMock
framework is interfering with it, but as for why this fixes it...
I can only guess.