Fix a bug where we were loading frameworks from the build directory
This is a fun one... xcodebuild build-for-testing adds the derivedData directory
to the DYLD_FRAMEWORK_PATH and DYLD_LIBRARY_PATH environment
variables so that if a framework is built but not added to the app bundle, the app
will still load in the simulator. Obviously, once the app is installed in an actual device
it won't load.

This commit clears up those variables so that the only frameworks available to the
app are the ones bundled with it. It also fixes the logic for detecting a crashed application
which was wrong (if the app crashed early enough, like when loading dynamic libraries).