Clean up the console output a little bit
This removes the [ADDITIONAL FAILURE] messages which are a bit misleading
because they mean there were additional asserts in the same run of the test which [ FAILED ].
It doesn't mean that the test was retried and it failed again. Note we still log everything in the
output directory.

Also only print the "Using %d out of %d processes" message when we hit the high watermark.