Simplify SimulatorMonitor and fix races (#236)
* Simplify SimulatorMonitor and fix races

SimulatorMonitor was trying to keep track of multiple things using a
single state variable which caused a bunch of problems. One of them
was a race between when the app finishes (the process exits) and the
output parser has finished parsing. We use a FIFO for grabbing the
output so it's buffered. This means that after the app has finished,
the parser is still parsing for short while (how short depends on how
fast the machine is).

* Simplify by using just one Enum.