Remove pipe for communication between app and bp
This fixes a bug where if the output of a failed test was too big,
then bp would deadlock with the app and think the app was hung.
It would essentially turn a test failure into a timeout.

Rather than worry about the corners of properly synchronizing over
a named pipe, I just have the app log to a file and wire the parser
to read from that file.