Merge exit code fixes from PR#432 and PR#430 to Xcode 11.2 branch (#435)
* make retry checks more resilient to prevent infinite retries (#432)

Also, log when retry count exceeded max retries.

Co-authored-by: Mansfield Mark <mmark@pinterest.com>

* Added a few more tests to mock test failure scenarios and fixed a few final Exit Status issues  (#430)

* Adding the ability to mock different test executions in Bluepill

* Add a few more tests and fixed the exit status issues

Changing the Exit Status to powers of 2 to make them mergeable.
Easy consolidation/aggregation of exit code.
In case of failure reporting, report all exit status that happened over multiple attempts, if any.
Prevent extra attempt, which does nothing, by eliminating the use of `hasRemainingTestsInContext` which is not accurate.

* Re-enabling some disabled/skipped BP tests and fixed a fragile packer test

A fragile packer test broke when the total number of tests changed. So, fixed
the testing method to fix the test.

Also, re-enabled a few tests in BluepillTests that were being skipped/disabled
for some reason.

* Renaming BPExitStatusTestsAllPassed to BPExitStatusAllTestsPassed

Renaming the all-good exit status from BPExitStatusTestsAllPassed
to BPExitStatusAllTestsPassed and incorporated comments. Tweaked
the tests a bit to reduce the regression testing duration.

Co-authored-by: Mansfield Mark <RainNapper@users.noreply.github.com>
Co-authored-by: Mansfield Mark <mmark@pinterest.com>