Add a flag to disable Xcode version check failure (#436)
* Add a flag to disable Xcode version check failure

For Xcode 10 and 11, many Xcode PRs to update bluepill have required simply
updating the default SDK and Xcode version. Throughout most releases of Xcode
11.4, bluepill has worked without a lot of major changes. At the speed that Apple
is moving on Xcode releases, it's a lot of effort to bump these: PRs,
branching, rebuilding, etc.

This flag makes it a bit easier for us to test new versions of Xcode
without having to create a build and PR for bluepill.

For now, it's defaulted, but perhaps we should make it default to on and print
a warning if it doesn't match. Additionally, we should consider dynamically
pulling in the runtime version

Co-authored-by: Ravi Mandala <rmandala@linkedin.com>