Do not retry crashed tests (#443)
Based on previous changes crashed tests are fatal and
would always return a non-zero exit code.

With this change, the crashed tests are NOT retried
to make sure the app crash errors are surfaced without confusion.
Also added new tests and fixed existing ones as per the new no-retry behavior.