Retry app crash tests and consider then non-fatal if they pass (#456)
Introducing a flag `retry-app-crash-tests` to retry tests that
might have caused app crashes. When this flag is turned on, the
execution will not fail if all app crash tests pass on retry.

The retries will honor the `error-retries` count, restarts the
sim before retrying the crashed test and makes the crash logs
available, like before.